recover the change files
