--- frontend/src/App.js
@@ -1,13 +1,0 @@
-import HomePage from './components/HomePage/HomePage.js';
-
-function App() {
-  return (
-    <div className="App">
-        <HomePage>
-
-        </HomePage>
-    </div>
-  );
-}
-
-export default App;
