+++ frontend/src/components/NaviBar/NaviBarStyles.js
